pyproject.toml
[tool.poetry]
name = "party-wall-flask"
version = "0.1.0"
description = ""
authors = ["15bubbles <15bubbles.15bubbles@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
Flask-RESTful = "^0.3.8"
flask-sqlalchemy = "^2.4.4"
flask-marshmallow = "^0.14.0"
gunicorn = "^20.0.4"
marshmallow-sqlalchemy = "^0.24.0"
psycopg2 = "^2.8.6"
Flask-Migrate = "^2.5.3"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pylint = "^2.6.0"
mypy = "^0.790"
isort = "^5.6.4"
pytest = "^6.1.1"
pytest-mock = "^3.3.1"
pytest-black = "^0.3.11"
pytest-pylint = "^0.17.0"
pytest-cov = "^2.10.1"
pytest-randomly = "^3.4.1"
pytest-sugar = "^0.9.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
